backend/requirements.txt
asgiref==3.8.1
astroid==3.3.5
colorama==0.4.6
coverage==7.6.8
dill==0.3.9
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
execnet==2.1.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
PyJWT==2.10.0
pylint==3.3.1
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
pytest-xdist==3.6.1
python-dotenv==1.0.1
sqlparse==0.5.2
tomlkit==0.13.2
tzdata==2024.2
